Correct handling of end of member blocks
There is now an optional limitation of block reading that allows the
length of the upcoming block to be queried before staring to read it. If
gz.BlockLimitted is not set, the behaviour matches the behaviour of gzip
from the standard library with the exception that headers are saved each
block.