Correct faulty interpretation of SAM spec
I was previously encoding the inflated size into the BC extra field,
this is not correct. Write the deflated length from a bytes.Buffer and
then copy that to the underlying writer.

Also need to hop over gzip member boundary during read. This needs to be
dealt with further when bgzf does parallel reading.