Perform initial compressBound check
The compressBound function is from zlib1.2.7, so need to keep an eye on
this in case the flate algorithms differ. We still check the length of
the compressed buffer before writing out, so errors will be caught, but
it would be nice to be able to guarantee that a filure won't occur
during a write.