Clean up uncovered chunk handling
The SAM spec really leaves something to be desired in terms of clarity.

The spec states: "Given a region [rbeg,rend), we only need to visit a
chunk whose end file offset is larger than the file offset of the 16kbp
window containing rbeg." If you wish to avoid unnecessary seeks to non-
covered bins then we check that the interval offset is valid (non-zero).
This means that we may miss covered bins unless we check, potentially
all, intervals in the region covered by the bin for the condition above,
not just the tile that covers the left end of the region.