Use ReadFull-ish
This makes sure a read gets the full slice or fails as for io.ReadFull.
io.ReadFull is not used because bgzf.Reader may return an io.EOF for a
zero-length []byte, and we make use of that to tighten the BAI indexing,
the first Read call outside the loop detects this io.EOF in cases where
len(p) == 0.