internal/index: fix overly aggressive optimisation
When all records in the first BGZF block are in the same bin, the tile
offset will be zero, so those records were incorrectly marked as not
used. Make the optimisation depend on having seen non-zero tiles
previously. This now depends on the fact that for any individual chunk,
the records must have been written in position order.