Add Cigar.Lengths()
When parsing SA tags or MD, we may also need to know the length often
the chimeric alignment or determine the length of the Cigar. With this
change, it is possible to do so without replicating code in Record.End()

report both ref and read consumption via Lengths()

update as per review

add test for N and fix bug it found. move tests.