csi: fix stats dummy bin handling
Note that the depth numbering in SAM and CSI disagree. The SAM spec
correctly states that the index is 6 deep, but CSI uses a default depth
of 5, while still produding the same bin limit value. Thus it appears
that CSI has an off-by-one entrenched in the documentation and code. I
suspect that this has come from the way the example bin calculation code
in the SAM spec (section 5.3) is formatted with 5 lines of code (the
zero bin being implicitly calculated with the fall through in reg2bin
and the zero assignment in reg2bins.

Long live -1-base counting!